+++ support/release-notes.md
@@ -1,2 +1,9 @@
+New:
+    - Data parameters from PHPUnit are now shown in the "Parameters" tab in the results pane
+    - `console` helper in PHPUnit to help debug tests while running Lode
+Changed:
+    - Improved parsing of Jest failure feedback
+    - Improved rendering of feedback text
+    - Always use panels for diff
 Fixed:
-    - Security fixes - #10 #11
+    - PHPUnit unable to run tests in separate processes